Refactor: test코드의 위치를 수정하였습니다.
