+++ src/test/java/baseball/RandomNumTest.java
@@ -1,8 +1,7 @@
-package study;
+package baseball;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import baseball.Application;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import java.util.List;
